requirements.txt
fastapi
pandas
joblib
uvicorn[standard]
numpy
scikit-learn
opencv-python-headless
pillow